timeless localCart

+++ src/components/DeleteCartButton.js
@@ -2,20 +2,40 @@
 import { X, XCircle, XLogo } from "@phosphor-icons/react"
 import axios from "axios"
 import { useRouter } from "next/navigation"
+import { useEffect, useState } from "react"
 
-const DeleteCartButton = ({ detailSlug, toast}) => {
+const DeleteCartButton = ({ detailSlug, toast, detailLocalSlug, updateLocalCart }) => {
+    const [authUser, setAuthUser] = useState(null)
     const router = useRouter()
+    useEffect(() => {
+        const fetchUser = async () => {
+            const { data } = await axios.get('/api/v1/cart/check');
+            setAuthUser(data.user)
+        }
+        fetchUser()
+    }, [])
     const handleDeleteCart = async () => {
-        const deleteCart = await axios.delete(`/api/v1/cart/${detailSlug}`)
-        if (deleteCart.data.status === 200) {
+        if (!authUser) {
+            const getCartItem = JSON.parse(localStorage.getItem("local-cart")) || []
+            const deleteCartItem = getCartItem.filter((item) => item.slug !== detailLocalSlug)
+            localStorage.setItem('local-cart', JSON.stringify(deleteCartItem))
             toast.loading("Tunggu...", { duration: 1000 })
             setTimeout(() => {
                 toast.success("Furniture terhapus dari keranjang", { icon: '🗑️' })
-                router.refresh()
+                updateLocalCart(deleteCartItem)
             }, 1000)
-            const getCartItem = JSON.parse(localStorage.getItem("cartItems")) || []
-            const deleteCartItem = getCartItem.filter((item) => item.slug !== detailSlug)
-            localStorage.setItem('cartItems', JSON.stringify(deleteCartItem))
+        } else {
+            const deleteCart = await axios.delete(`/api/v1/cart/${detailSlug}`)
+            if (deleteCart.data.status === 200) {
+                toast.loading("Tunggu...", { duration: 1000 })
+                setTimeout(() => {
+                    toast.success("Furniture terhapus dari keranjang", { icon: '🗑️' })
+                    router.refresh()
+                }, 1000)
+                const getCartItem = JSON.parse(localStorage.getItem("local-cart")) || []
+                const deleteCartItem = getCartItem.filter((item) => item.slug !== detailSlug)
+                localStorage.setItem('local-cart', JSON.stringify(deleteCartItem))
+            }
         }
     }
     return (
